Karaoke Song List
...plus some rearranging of the code so that each view is in its own file.
