+++ lib/src/models/user.dart
@@ -1,0 +1,6 @@
+class User {
+  const User({ this.name, this.email });
+
+  final String name;
+  final String email;
+}
