+++ pubspec.yaml
@@ -22,3 +22,4 @@ flutter:
    - images/deck_8.png
    - images/deck_9.png
    - images/deck_10.png
+   - resources/JoCoKaraokeSongCatalog.txt
