+++ test/calendar_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:cruisemonkey/main.dart';
-import 'package:cruisemonkey/models.dart';
+import 'package:cruisemonkey/src/models/calendar.dart';
 
 import 'mocks.dart';
